+++ okapi-operation-macro/src/operation/cookie.rs
@@ -39,14 +39,14 @@ impl ToTokens for Cookie {
         tokens.extend(quote! {
             okapi::openapi3::Parameter {
                 name: #name.into(),
-                location: "Cookie".into(),
+                location: "cookie".into(),
                 description: #description,
                 required: #required,
                 deprecated: #deprecated,
                 allow_empty_value: #allow_empty_values,
                 value: {
                     okapi::openapi3::ParameterValue::Schema {
-                        style: #style,
+                        style: Some(#style),
                         explode: #explode,
                         allow_reserved: #allow_reserved,
                         schema: components.schema_for::<#ty>(),